product and order - all done

+++ src/main/java/com/example/ecommerce/repository/OrderRepository.java
@@ -3,4 +3,5 @@ package com.example.ecommerce.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.ecommerce.model.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {}
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+}
